Add average formant value finding based on average values for each vowel
Found example of displaying waveform, this will probably be implemented for pronunciation exercises.
Recording has clipping issues that I might now what to do about.
If the formant values given to me by my supervisor are averages on the whole (and not just for males), then sex plays no role in profile creation.
Formant finding is now in a separate FormantUtils class.